# contracts/openapi/v1-engine.json
{
  "components": {
    "schemas": {
...
      "ApprovalRequest": {
        "properties": {
          "command_id": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Command Id",
            "type": "string"
          },
          "digest": {
            "title": "Digest",
            "type": "string"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          }
        },
        "required": [
          "command_id",
          "digest",
          "expected_version"
        ],
        "title": "ApprovalRequest",
        "type": "object"
      },
...
      "ComparisonLimits": {
        "properties": {
          "max_attempts": {
            "title": "Max Attempts",
            "type": "integer"
          },
          "max_rows": {
            "title": "Max Rows",
            "type": "integer"
          }
        },
        "required": [
          "max_rows",
          "max_attempts"
        ],
        "title": "ComparisonLimits",
        "type": "object"
      },
      "ComparisonList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ComparisonProposal"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "ComparisonList",
        "type": "object"
      },
      "ComparisonProposal": {
        "properties": {
          "artifact_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Id"
          },
          "digest": {
            "title": "Digest",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "expected_version": {
            "title": "Expected Version",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "left": {
            "$ref": "#/components/schemas/ComparisonSource"
          },
          "limits": {
            "$ref": "#/components/schemas/ComparisonLimits"
          },
          "report_csv": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report Csv"
          },
          "report_markdown": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report Markdown"
          },
          "right": {
            "$ref": "#/components/schemas/ComparisonSource"
          },
          "status": {
            "enum": [
              "pending_approval",
              "queued",
              "running",
              "completed",
              "failed",
              "blocked"
            ],
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "tool_version": {
            "title": "Tool Version",
            "type": "string"
          },
          "work_id": {
            "title": "Work Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "work_id",
          "digest",
          "expected_version",
          "tool_version",
          "left",
          "right",
          "limits"
        ],
        "title": "ComparisonProposal",
        "type": "object"
      },
      "ComparisonSource": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "title",
          "sha256",
          "version"
        ],
        "title": "ComparisonSource",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProposalRequest": {
        "properties": {
          "command_id": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Command Id",
            "type": "string"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "left_material_id": {
            "title": "Left Material Id",
            "type": "string"
          },
          "max_rows": {
            "default": 10000,
            "maximum": 10000.0,
            "minimum": 1.0,
            "title": "Max Rows",
            "type": "integer"
          },
          "right_material_id": {
            "title": "Right Material Id",
            "type": "string"
          }
        },
        "required": [
          "command_id",
          "left_material_id",
          "right_material_id",
          "expected_version"
        ],
        "title": "ProposalRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Local Homun 2 engine — domain F2 over SQLite + HTTP.",
    "title": "Homun Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/workspaces/{workspace_id}/works/{work_id}/price-comparisons": {
      "get": {
        "operationId": "get_proposals_v1_workspaces__workspace_id__works__work_id__price_comparisons_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "work_id",
            "required": true,
            "schema": {
              "title": "Work Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-homun-actor-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Homun-Actor-Id"
            }
          },
          {
            "in": "header",
            "name": "x-homun-actor-name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Homun-Actor-Name"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComparisonList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Proposals",
        "tags": [
          "price-comparisons"
        ]
      },
      "post": {
        "operationId": "create_proposal_v1_workspaces__workspace_id__works__work_id__price_comparisons_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "work_id",
            "required": true,
            "schema": {
              "title": "Work Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-homun-actor-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Homun-Actor-Id"
            }
          },
          {
            "in": "header",
            "name": "x-homun-actor-name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Homun-Actor-Name"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComparisonProposal"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Proposal",
        "tags": [
          "price-comparisons"
        ]
      }
    },
    "/v1/workspaces/{workspace_id}/works/{work_id}/price-comparisons/{proposal_id}/approve": {
      "post": {
        "operationId": "approve_proposal_v1_workspaces__workspace_id__works__work_id__price_comparisons__proposal_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "work_id",
            "required": true,
            "schema": {
              "title": "Work Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-homun-actor-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Homun-Actor-Id"
            }
          },
          {
            "in": "header",
            "name": "x-homun-actor-name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Homun-Actor-Name"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComparisonProposal"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Proposal",
        "tags": [
          "price-comparisons"
        ]
      }
    },
...
  }
}